+++ backend/runtimes/local/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @cat-factory/local-server
 
+## 0.34.2
+
+### Patch Changes
+
+- Updated dependencies [d4d4cbc]
+  - @cat-factory/server@0.62.1
+  - @cat-factory/integrations@0.51.1
+  - @cat-factory/node-server@0.54.1
+  - @cat-factory/orchestration@0.52.1
+
 ## 0.34.1
 
 ### Patch Changes